Remove old swap code
