+++ src/cow-react/modules/swap/containers/ConfirmSwapModalSetup/index.tsx
@@ -5,7 +5,6 @@ import { Percent } from '@uniswap/sdk-core'
 import { Field } from 'state/swap/actions'
 import { useAtomValue } from 'jotai/utils'
 import { swapConfirmAtom } from '@cow/modules/swap/state/swapConfirmAtom'
-import { HandleSwapCallback } from '@cow/modules/swap/hooks/useHandleSwap'
 import { useSwapConfirmManager } from '@cow/modules/swap/hooks/useSwapConfirmManager'
 import TransactionConfirmationModal from 'components/TransactionConfirmationModal'
 import { useSwapActionHandlers } from 'state/swap/hooks'
@@ -13,6 +12,7 @@ import { useModalIsOpen } from 'state/application/hooks'
 import { ApplicationModal } from 'state/application/reducer'
 import { useCloseModals } from 'state/application/hooks'
 import { transactionConfirmAtom } from '@cow/modules/swap/state/transactionConfirmAtom'
+import { HandleSwapCallback } from '@cow/modules/swap/pure/SwapButtons'
 
 export interface ConfirmSwapModalSetupProps {
   trade: TradeGp | undefined
