+++ src/cow-react/modules/swap/containers/FeesDiscount/index.tsx
@@ -6,7 +6,7 @@ import { AutoRow } from 'components/Row'
 import { Trans } from '@lingui/macro'
 import { MouseoverTooltipContent } from 'components/Tooltip'
 import { SUBSIDY_INFO_MESSAGE } from 'components/CowSubsidyModal/constants'
-import { StyledInfo } from '@cow/pages/Swap/styleds'
+import { StyledInfo } from '../../pure/styled'
 
 interface FeesDiscountProps extends BoxProps {
   theme: DefaultTheme
