+++ src/cow-react/modules/swap/hooks/useSwapFlowContext.ts
@@ -18,11 +18,11 @@ import { PostOrderParams } from 'utils/trade'
 import { Currency, CurrencyAmount, TradeType } from '@uniswap/sdk-core'
 import { OrderKind } from '@cowprotocol/contracts'
 import { NATIVE_CURRENCY_BUY_TOKEN } from 'constants/index'
-import { MAX_VALID_TO_EPOCH } from 'hooks/useSwapCallback'
 import { useUserTransactionTTL } from 'state/user/hooks'
 import { useGP2SettlementContract } from 'hooks/useContract'
 import { useUpdateAtom } from 'jotai/utils'
 import { addAppDataToUploadQueueAtom } from 'state/appData/atoms'
+import { calculateValidTo } from '@cow/common/utils/calculateValidTo'
 
 const _computeInputAmountForSignature = (params: {
   input: CurrencyAmount<Currency>
@@ -42,15 +42,6 @@ const _computeInputAmountForSignature = (params: {
     // User BUYING? POST inputAmount as amount with no fee
     return inputWithSlippage.subtract(fee)
   }
-}
-
-function calculateValidTo(deadline: number): number {
-  // Need the timestamp in seconds
-  const now = Date.now() / 1000
-  // Must be an integer
-  const validTo = Math.floor(deadline + now)
-  // Should not be greater than what the contract supports
-  return Math.min(validTo, MAX_VALID_TO_EPOCH)
 }
 
 export function useSwapFlowContext(): SwapFlowContext | null {
