+++ src/cow-react/modules/swap/pure/Row/RowReceivedAfterSlippageContent/index.tsx
@@ -1,6 +1,6 @@
 import { Trans } from '@lingui/macro'
 import { CurrencyAmount, Currency, TradeType } from '@uniswap/sdk-core'
-import { StyledInfo } from '@cow/pages/Swap/styleds'
+import { StyledInfo } from '../../styled'
 import { RowReceivedAfterSlippageProps } from '@cow/modules/swap/containers/RowReceivedAfterSlippage'
 
 import { RowFixed } from 'components/Row'
