+++ src/cow-react/modules/swap/pure/Row/RowSlippageContent/index.tsx
@@ -4,7 +4,7 @@ import { Trans } from '@lingui/macro'
 import { RowFixed } from 'components/Row'
 import { MouseoverTooltipContent } from 'components/Tooltip'
 import { INPUT_OUTPUT_EXPLANATION } from 'constants/index'
-import { StyledInfo } from '@cow/pages/Swap/styleds'
+import { StyledInfo } from '../../styled'
 import { RowSlippageProps } from '@cow/modules/swap/containers/RowSlippage'
 import { StyledRowBetween, TextWrapper } from '@cow/modules/swap/pure/Row/styled'
 import { RowStyleProps } from '@cow/modules/swap/pure/Row/typings'
