+++ src/custom/state/user/hooks/index.ts
@@ -7,9 +7,9 @@ import {
   removeAllFavouriteTokens,
   initFavouriteTokens,
 } from 'state/user/reducer'
-import { calculateValidTo } from 'hooks/useSwapCallback'
 import { useUserTransactionTTL, serializeToken, deserializeToken } from '@src/state/user/hooks'
 import { useWeb3React } from '@web3-react/core'
+import { calculateValidTo } from '@cow/common/utils/calculateValidTo'
 
 export * from '@src/state/user/hooks'
